+++ common/src/main/java/com/evandev/redomesticate/mixin/TamableAnimalMixin.java
@@ -8,6 +8,7 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(TamableAnimal.class)
 public abstract class TamableAnimalMixin {
@@ -22,6 +23,13 @@ public abstract class TamableAnimalMixin {
             if (pOrderedToSit != shouldSit) {
                 ci.cancel();
             }
+        }
+    }
+
+    @Inject(method = "shouldTryTeleportToOwner", at = @At("HEAD"), cancellable = true)
+    private void redomesticate$disableTeleportation(CallbackInfoReturnable<Boolean> cir) {
+        if (ModConfig.get().disablePetTeleportation) {
+            cir.setReturnValue(false);
         }
     }
 }