+++ fabric/src/main/java/com/evandev/redomesticate/client/RedomesticateFabricClient.java
@@ -7,6 +7,7 @@ import com.evandev.redomesticate.client.registry.OreColorRegistry;
 import com.evandev.redomesticate.client.render.*;
 import com.evandev.redomesticate.content.entity.HighlightedBlockEntity;
 import com.evandev.redomesticate.content.item.DeedOfOwnershipItem;
+import com.evandev.redomesticate.content.item.FeatherOnAStickItem;
 import com.evandev.redomesticate.network.FabricNetworking;
 import com.evandev.redomesticate.registry.ModEntities;
 import com.evandev.redomesticate.registry.ModItems;
@@ -23,6 +24,7 @@ import net.minecraft.client.renderer.item.ItemProperties;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.player.Player;
 
 public class RedomesticateFabricClient implements ClientModInitializer {
 
@@ -59,6 +61,21 @@ public class RedomesticateFabricClient implements ClientModInitializer {
                 ResourceLocation.fromNamespaceAndPath(Constants.MOD_ID, "bound"),
                 (stack, level, entity, seed) -> DeedOfOwnershipItem.isBound(stack) ? 1.0F : 0.0F
         );
+
+        ItemProperties.register(ModItems.FEATHER_ON_A_STICK.get(), ResourceLocation.withDefaultNamespace("cast"), (stack, level, entity, seed) -> {
+            if (entity == null) {
+                return 0.0F;
+            } else {
+                boolean isMainHand = entity.getMainHandItem() == stack;
+                boolean isOffHand = entity.getOffhandItem() == stack;
+
+                if (entity.getMainHandItem().getItem() instanceof FeatherOnAStickItem) {
+                    isOffHand = false;
+                }
+
+                return (isMainHand || isOffHand) && entity instanceof Player player && player.fishing != null ? 1.0F : 0.0F;
+            }
+        });
 
         LivingEntityFeatureRendererRegistrationCallback.EVENT.register((entityType, entityRenderer, registrationHelper, context) -> {
             if (LayerManager.canApply(entityType) && entityType != EntityType.ENDER_DRAGON) {