+++ neoforge/src/main/java/com/evandev/redomesticate/datagen/providers/ModDatapackProvider.java
@@ -19,6 +19,5 @@ public class ModDatapackProvider extends DatapackBuiltinEntriesProvider {
 
     public ModDatapackProvider(PackOutput output, CompletableFuture<HolderLookup.Provider> registries) {
         super(output, registries, BUILDER, Set.of(Constants.MOD_ID));
-
     }
 }